Create a state and a module to create images using import tasks
In Glance V2 it is possible to import an image from a remote
location only using a tasks workflow. [1] As tasks are not
supported in salt.modules.glance I had to create a new module
glanceng that extends functionality of the glance module.
That is supposed to be a temporary solution until we move to a
SaltStack version that has support of Glance V2 tasks.

Added the `get_image_owner_id` function to the glanceng module
that is used for mining owner of a created image in Octavia
stare.

Added a new state image_import to _states/glanceng that creates
a task with type `import` and the verifies that the task succeeded
and an image was created.

Updated glance/client.sls to use image_import state and extended
the list of parameters that can be passed to that state.

[1] https://wiki.openstack.org/wiki/Glance-tasks-import

Change-Id: Ica8d02ed4a1653a74ac87ff5ee0efbe5d19feed0